test(user): scaffold Vitest and add helper for user creation

+++ scripts/test-helpers/tests.ts
@@ -1,13 +1,20 @@
-import { createTestUser } from './test-create-user';
+import { createTestUser } from './create-test-user';
 import { deleteTestUser } from './test-delete-user';
 import { createTestThread } from './test-create-thread';
 import { viewTables } from './test-view-tables';
 
-async function run() {
+const cleanup = false;
+
+async function run(cleanup: boolean = false) {
+  if (cleanup) {
+    await deleteTestUser();
+    await viewTables();
+    return;
+  }
   // await createTestUser();
   // await createTestThread();
-  await deleteTestUser();
+
   await viewTables();
 }
 
-run();
+run(cleanup);
